Configure note headline as slug
This commit configures the first headline of a note as the slug of the ox-hugo
file. As a result, there's no need to use that headline as the name of the
ox-hugo file anymore.

This also means that the ox-hugo note file now uses the note date as its
filename, just as the orginal note file.